refactor : core폴더 생성
