+++ src/Component/index.js
@@ -1,3 +1,3 @@
-export { default as Component } from "./Component.js";
+export { default as Component } from "../core/Component.js";
 export { default as DocumentTreeComponent } from "./DocumentTree.component.js";
 export { default as EditorComponent } from "./Editor.component.js";
